// src/app/modules/campaign-setting/setting-goal/models/CommunicationRecordDepositListModel.ts
export interface CommunicationRecordDepositListModel{
    goalTypeCommunicationRecordDepositId: number
    sequenceId: number
    sequenceName: string
    goalTypeId : number
    goalTypeName: string
    campaignSettingId: number
    goalTypeDataSourceName: string
    goalTypePeriodName: string
    //goalType: string
    createdBy: string
    createdDate?: string | any
    updatedBy: string
    updatedDate?: string | any
    goalTypeGuid?: string | any   
}